# openapi.yaml
openapi: 3.0.3
info:
  title: Redstick Ventures Command Center API
  description: |
    Complete API for the Redstick Ventures Command Center platform.
    
    ## Authentication
    All API requests require a JWT token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```
    
    ## Rate Limiting
    - Authenticated: 1000 requests/minute
    - Unauthenticated: 20 requests/minute
    
    ## Base URLs
    - Production: https://api.redstick.vc
    - Staging: https://api-staging.redstick.vc
    - Development: http://localhost:3000/api
  version: 1.0.0
  contact:
    name: Redstick API Support
    email: api-support@redstick.vc
    url: https://docs.redstick.vc
  license:
    name: Proprietary
    url: https://redstick.vc/terms
servers:
  - url: https://api.redstick.vc/api/v1
    description: Production
  - url: https://api-staging.redstick.vc/api/v1
    description: Staging
  - url: http://localhost:3000/api/v1
    description: Development
security:
  - bearerAuth: []
paths:
...
  /agents/{id}:
    get:
      tags:
        - Agents
      summary: Get agent details
      parameters:
        - $ref: '#/components/parameters/AgentIdParam'
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentDetailResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login
  parameters:
...
    AgentIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Agent ID
...
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        data:
          type: object

    # Pagination
...
    Agent:
      type: object
      required:
        - id
        - name
        - status
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        icon:
          type: string
        status:
          type: string
          enum: [active, inactive]
        capabilities:
          type: array
          items:
            type: string
        inputSchema:
          type: object
        outputSchema:
          type: object
        avgExecutionTime:
          type: integer
          description: Average execution time in seconds
        successRate:
          type: number
          description: Success rate between 0 and 1
        createdAt:
          type: string
          format: date-time
    AgentResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Agent'
...
    AgentDetailResponse:
      allOf:
        - $ref: '#/components/schemas/AgentResponse'
...