# src/app/api/agents/[id]/route.ts
import { NextRequest, NextResponse } from 'next/server'
import { Agent, UpdateAgentInput } from '@/types/agent'

// Mock data store - In production, this would be a database
// Shared with the parent route
let agents: Agent[] = [
  {
    id: 'agent-001',
    name: 'Deal Screener Pro',
    description: 'Automatically screens incoming deal flow based on investment criteria, team quality, and market size.',
    type: 'SCREENING',
    status: 'ACTIVE',
    config: {
      minTeamSize: 2,
      minMarketSize: 10000000,
      industries: ['SaaS', 'AI', 'Fintech'],
      autoReject: true
    },
    lastRun: '2026-03-27T10:30:00Z',
    successRate: 0.94,
    totalRuns: 156,
    tokenUsage: 456000
  },
  {
    id: 'agent-002',
    name: 'Market Research Analyst',
    description: 'Conducts comprehensive market research including TAM/SAM/SOM analysis, competitive landscape, and trend identification.',
    type: 'RESEARCH',
    status: 'ACTIVE',
    config: {
      sources: ['Crunchbase', 'PitchBook', 'LinkedIn', 'Industry Reports'],
      deepDive: true,
      reportFormat: 'standard'
    },
    lastRun: '2026-03-27T14:15:00Z',
    successRate: 0.89,
    totalRuns: 89,
    tokenUsage: 892000
  },
  {
    id: 'agent-003',
    name: 'Portfolio Monitor',
    description: 'Tracks portfolio company performance, monitors KPIs, and alerts on potential issues or milestones.',
    type: 'PORTFOLIO',
    status: 'ACTIVE',
    config: {
      metrics: ['ARR', 'Burn Rate', 'Headcount', 'NRR'],
      alertThresholds: {
        burnRate: 6,
        runway: 12
      },
      checkFrequency: 'daily'
    },
    lastRun: '2026-03-28T08:00:00Z',
    successRate: 0.97,
    totalRuns: 312,
    tokenUsage: 234000
  },
  {
    id: 'agent-004',
    name: 'Outreach Assistant',
    description: 'Drafts personalized outreach emails to founders and manages follow-up sequences.',
    type: 'OUTREACH',
    status: 'INACTIVE',
    config: {
      tone: 'professional',
      maxFollowUps: 3,
      personalizationLevel: 'high',
      templates: ['introduction', 'follow_up', 'pass']
    },
    lastRun: '2026-03-20T16:45:00Z',
    successRate: 0.76,
    totalRuns: 45,
    tokenUsage: 178000
  },
  {
    id: 'agent-005',
    name: 'Due Diligence Assistant',
    description: 'Assists with due diligence by analyzing legal documents, financial statements, and reference checks.',
    type: 'DILIGENCE',
    status: 'MAINTENANCE',
    config: {
      checkCategories: ['legal', 'financial', 'technical', 'commercial'],
      riskThreshold: 'medium',
      generateReport: true
    },
    lastRun: '2026-03-25T11:20:00Z',
    successRate: 0.88,
    totalRuns: 23,
    tokenUsage: 567000
  },
  {
    id: 'agent-006',
    name: 'LP Report Generator',
    description: 'Generates quarterly LP reports with portfolio performance, deployment updates, and market insights.',
    type: 'REPORTING',
    status: 'ACTIVE',
    config: {
      sections: ['overview', 'performance', 'deployment', 'exits', 'pipeline'],
      format: 'pdf',
      includeBenchmarks: true
    },
    lastRun: '2026-03-01T09:00:00Z',
    successRate: 0.92,
    totalRuns: 12,
    tokenUsage: 345000
  }
]

// Helper function to find agent by ID
function findAgentById(id: string): Agent | undefined {
  return agents.find(agent => agent.id === id)
}

// GET /api/agents/[id] - Get agent details
export async function GET(
  request: NextRequest,
  { params }: { params: Promise<{ id: string }> }
): Promise<NextResponse> {
  try {
    const { id } = await params
    
    if (!id) {
      return NextResponse.json({
        success: false,
        error: 'Agent ID is required'
      }, { status: 400 })
    }
    
    const agent = findAgentById(id)
    
    if (!agent) {
      return NextResponse.json({
        success: false,
        error: 'Agent not found'
      }, { status: 404 })
    }
    
    return NextResponse.json({
      success: true,
      data: agent
    }, { status: 200 })
  } catch (error) {
    console.error('Error fetching agent:', error)
    return NextResponse.json({
      success: false,
      error: 'Failed to fetch agent'
    }, { status: 500 })
  }
}

// PATCH /api/agents/[id] - Update agent
export async function PATCH(
  request: NextRequest,
  { params }: { params: Promise<{ id: string }> }
): Promise<NextResponse> {
  try {
    const { id } = await params
    
    if (!id) {
      return NextResponse.json({
        success: false,
        error: 'Agent ID is required'
      }, { status: 400 })
    }
    
    const agentIndex = agents.findIndex(agent => agent.id === id)
    
    if (agentIndex === -1) {
      return NextResponse.json({
        success: false,
        error: 'Agent not found'
      }, { status: 404 })
    }
    
    const body: UpdateAgentInput = await request.json()
    
    // Validate status if provided
    if (body.status && !['ACTIVE', 'INACTIVE', 'ERROR', 'MAINTENANCE'].includes(body.status)) {
      return NextResponse.json({
        success: false,
        error: 'Invalid status. Must be ACTIVE, INACTIVE, ERROR, or MAINTENANCE'
      }, { status: 400 })
    }
    
    // Validate name if provided
    if (body.name !== undefined && (typeof body.name !== 'string' || body.name.trim().length === 0)) {
      return NextResponse.json({
        success: false,
        error: 'Name must be a non-empty string'
      }, { status: 400 })
    }
    
    // Validate description if provided
    if (body.description !== undefined && typeof body.description !== 'string') {
      return NextResponse.json({
        success: false,
        error: 'Description must be a string'
      }, { status: 400 })
    }
    
    // Update agent fields
    const updatedAgent: Agent = {
      ...agents[agentIndex],
      ...(body.name !== undefined && { name: body.name.trim() }),
      ...(body.description !== undefined && { description: body.description.trim() }),
      ...(body.status !== undefined && { status: body.status }),
      ...(body.config !== undefined && { config: { ...agents[agentIndex].config, ...body.config } })
    }
    
    agents[agentIndex] = updatedAgent
    
    return NextResponse.json({
      success: true,
      data: updatedAgent,
      message: 'Agent updated successfully'
    }, { status: 200 })
  } catch (error) {
    console.error('Error updating agent:', error)
    return NextResponse.json({
      success: false,
      error: 'Failed to update agent'
    }, { status: 500 })
  }
}

// DELETE /api/agents/[id] - Delete agent
export async function DELETE(
  request: NextRequest,
  { params }: { params: Promise<{ id: string }> }
): Promise<NextResponse> {
  try {
    const { id } = await params
    
    if (!id) {
      return NextResponse.json({
        success: false,
        error: 'Agent ID is required'
      }, { status: 400 })
    }
    
    const agentIndex = agents.findIndex(agent => agent.id === id)
    
    if (agentIndex === -1) {
      return NextResponse.json({
        success: false,
        error: 'Agent not found'
      }, { status: 404 })
    }
    
    // Prevent deletion of active agents (optional business rule)
    if (agents[agentIndex].status === 'ACTIVE') {
      return NextResponse.json({
        success: false,
        error: 'Cannot delete an active agent. Please deactivate it first.'
      }, { status: 409 })
    }
    
    const deletedAgent = agents[agentIndex]
    agents.splice(agentIndex, 1)
    
    return NextResponse.json({
      success: true,
      data: { id: deletedAgent.id },
      message: 'Agent deleted successfully'
    }, { status: 200 })
  } catch (error) {
    console.error('Error deleting agent:', error)
    return NextResponse.json({
      success: false,
      error: 'Failed to delete agent'
    }, { status: 500 })
  }
}